custom-model/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML Core
torch>=2.1.0
transformers>=4.36.0
diffusers>=0.24.0
accelerate>=0.25.0
safetensors>=0.4.0

# SDXL & ControlNet
controlnet-aux>=0.0.7
ip-adapter>=0.1.0

# Segmentation
segment-anything>=1.0

# Agent Framework
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Lineage & Royalties
networkx>=3.2.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Storage
google-cloud-storage>=2.13.0

# Training
peft>=0.7.0
datasets>=2.15.0
wandb>=0.16.0

# Evaluation
torchmetrics>=1.2.0
lpips>=0.1.4

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
